+++ src/song.rs
@@ -1,5 +1,5 @@
 pub struct Song {
-    pub start: u32,
+    pub start: u64,
     pub index: usize,
     pub size: usize,
 }
